pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "lhd"
version = "0.1.0"
[tool.poetry.dependencies]
python = "^3.7"
bs4 = "^0.0.1"
favicon = "^0.7.0"
pip = "^20.1"
pip-upgrader = "^1.4"
psycopg2-binary = "^2.8"
flask = "^1.1"
flask-sqlalchemy = "^2.4"
sklearn = "^0.0.0"
sqlalchemy = "^1.3"
[tool.poetry.dev-dependencies]
